+++ client/src/components/EpicEntryItem/style.ts
@@ -27,25 +27,6 @@ const S = {
     transition: opacity 0.1s ease;
     opacity: ${({ showDraggable }) => (showDraggable ? 1 : 0)};
   `,
-  Form: styled.form`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    min-width: 350px;
-    padding: 16px;
-  `,
-  Input: styled.input`
-    width: 100%;
-    padding: 8px 16px;
-    margin-right: 16px;
-
-    background-color: ${({ theme }) => theme.color.gray100};
-    color: ${({ theme }) => theme.color.gray400};
-    border-radius: 8px;
-
-    font: ${({ theme }) => theme.font.body_regular};
-  `,
 };
 
 export default S;
